feat: introduce Plugins API, fix tests
