+++ src/parsers/OctetStream.js
@@ -2,6 +2,11 @@
 
 const { PassThrough } = require('stream');
 
-class OctetStreamParser extends PassThrough {}
+class OctetStreamParser extends PassThrough {
+  constructor(options = {}) {
+    super();
+    this.globalOptions = { ...options };
+  }
+}
 
 module.exports = OctetStreamParser;
