+++ test/unit/test-incoming-form.js
@@ -17,60 +17,64 @@ let form;
       form = name === 'formidable' ? mod.formidable() : new mod[name]();
     },
 
-    [`${name}#_fileName with regular characters`]: () => {
+    [`${name}#_getFileName with regular characters`]: () => {
       const filename = 'foo.txt';
 
-      assert.equal(form._fileName(makeHeader(filename)), 'foo.txt');
+      assert.strictEqual(form._getFileName(makeHeader(filename)), 'foo.txt');
     },
 
-    [`${name}#_fileName with unescaped quote`]: () => {
+    [`${name}#_getFileName with unescaped quote`]: () => {
       const filename = 'my".txt';
 
-      assert.equal(form._fileName(makeHeader(filename)), 'my".txt');
+      assert.strictEqual(form._getFileName(makeHeader(filename)), 'my".txt');
     },
 
-    [`${name}#_fileName with escaped quote`]: () => {
+    [`${name}#_getFileName with escaped quote`]: () => {
       const filename = 'my%22.txt';
 
-      assert.equal(form._fileName(makeHeader(filename)), 'my".txt');
+      assert.strictEqual(form._getFileName(makeHeader(filename)), 'my".txt');
     },
 
-    [`${name}#_fileName with bad quote and additional sub-header`]: () => {
+    [`${name}#_getFileName with bad quote and additional sub-header`]: () => {
       const filename = 'my".txt';
 
       const header = `${makeHeader(filename)}; foo="bar"`;
-      assert.equal(form._fileName(header), filename);
+      assert.strictEqual(form._getFileName(header), filename);
     },
 
-    [`${name}#_fileName with semicolon`]: () => {
+    [`${name}#_getFileName with semicolon`]: () => {
       const filename = 'my;.txt';
 
-      assert.equal(form._fileName(makeHeader(filename)), 'my;.txt');
+      assert.strictEqual(form._getFileName(makeHeader(filename)), 'my;.txt');
     },
 
-    [`${name}#_fileName with utf8 character`]: () => {
+    [`${name}#_getFileName with utf8 character`]: () => {
       const filename = 'my&#9731;.txt';
 
-      assert.equal(form._fileName(makeHeader(filename)), 'my☃.txt');
+      assert.strictEqual(form._getFileName(makeHeader(filename)), 'my☃.txt');
     },
 
     [`${name}#_uploadPath strips harmful characters from extension when keepExtensions`]: () => {
-      form.keepExtensions = true;
+      const opts = {
+        keepExtensions: true,
+      };
+
+      form = name === 'formidable' ? mod.formidable(opts) : new mod[name](opts);
 
       let ext = path.extname(form._uploadPath('fine.jpg?foo=bar'));
-      assert.equal(ext, '.jpg');
+      assert.strictEqual(ext, '.jpg');
 
       ext = path.extname(form._uploadPath('fine?foo=bar'));
-      assert.equal(ext, '');
+      assert.strictEqual(ext, '');
 
       ext = path.extname(form._uploadPath('super.cr2+dsad'));
-      assert.equal(ext, '.cr2');
+      assert.strictEqual(ext, '.cr2');
 
       ext = path.extname(form._uploadPath('super.bar'));
-      assert.equal(ext, '.bar');
+      assert.strictEqual(ext, '.bar');
 
       ext = path.extname(form._uploadPath('file.aAa'));
-      assert.equal(ext, '.aAa');
+      assert.strictEqual(ext, '.aAa');
     },
 
     [`${name}#_Array parameters support`]: () => {
@@ -78,11 +82,15 @@ let form;
       form = name === 'formidable' ? mod.formidable(opts) : new mod[name](opts);
 
       const req = new Request();
-      req.headers = 'content-type: json; content-length:8';
+      req.headers = {
+        'content-type': 'json',
+        'content-length': 8,
+      };
+
       form.parse(req, (error, fields) => {
-        assert.equal(Array.isArray(fields.a), true);
-        assert.equal(fields.a[0], 1);
-        assert.equal(fields.a[1], 2);
+        assert.strictEqual(Array.isArray(fields.a), true);
+        assert.strictEqual(fields.a[0], 1);
+        assert.strictEqual(fields.a[1], 2);
       });
       form.emit('field', 'a[]', 1);
       form.emit('field', 'a[]', 2);
